Separate smic definitions from input builder
