+++ openfisca_france_data/erfs_fpr/input_data_builder/step_04_famille.py
@@ -8,7 +8,7 @@ import pandas as pd
 
 from openfisca_survey_manager.temporary import temporary_store_decorator  # type: ignore
 from openfisca_france_data.utils import assert_dtype
-from openfisca_france_data.common import smic_horaire_brut
+from openfisca_france_data.smic import smic_horaire_brut
 
 
 log = logging.getLogger(__name__)
